// web/src/shared/api/generated/RequestLogListResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ActiveRequestLogResponse } from "./ActiveRequestLogResponse";
import type { RequestLogFilterOptionsResponse } from "./RequestLogFilterOptionsResponse";
import type { RequestLogResponse } from "./RequestLogResponse";
import type { RequestTelemetryResponse } from "./RequestTelemetryResponse";

export type RequestLogListResponse = { active_items: Array<ActiveRequestLogResponse>, active_total: number, items: Array<RequestLogResponse>, next_cursor: string | null, has_more: boolean, telemetry: RequestTelemetryResponse, filter_options: RequestLogFilterOptionsResponse, };