// web/src/shared/api/generated/RequestLogDetailResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RequestAttemptResponse } from "./RequestAttemptResponse";
import type { RequestLogResponse } from "./RequestLogResponse";
import type { RequestTelemetryResponse } from "./RequestTelemetryResponse";

export type RequestLogDetailResponse = { request: RequestLogResponse, attempts: Array<RequestAttemptResponse>, telemetry: RequestTelemetryResponse, };